Refactor iqk: Factor out GEMM for 1-bit quants (ABX2/AVX512)
