+++ ggml/src/CMakeLists.txt
@@ -264,6 +264,7 @@ if (GGML_IQK_MUL_MAT)
                             iqk/iqk_gemm_kquants.cpp
                             iqk/iqk_gemm_iquants.cpp
                             iqk/iqk_gemm_iqk_quants.cpp
+                            iqk/iqk_gemm_1bit.cpp
                             iqk/iqk_gemm_legacy_quants.cpp)
     set(GGML_HEADERS_IQK_MM iqk/iqk_mul_mat.h
                             iqk/iqk_flash_impl.h
@@ -271,6 +272,7 @@ if (GGML_IQK_MUL_MAT)
                             iqk/iqk_gemm_kquants.h
                             iqk/iqk_gemm_iquants.h
                             iqk/iqk_gemm_iqk_quants.h
+                            iqk/iqk_gemm_1bit.h
                             iqk/iqk_gemm_legacy_quants.h)
     if (GGML_IQK_FLASH_ATTENTION)
         message(STATUS "Enabling IQK Flash Attention kernels")
